// source/telemetry/generated/main.ts
/* This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/. */

// AUTOGENERATED BY glean_parser. DO NOT EDIT. DO NOT COMMIT.

import EventMetricType from "@mozilla/glean/private/metrics/event";

/**
 * Event sent after user clicks submit on onboarding page
 *
 * Generated from `main.onboarding_completed`.
 */
export const onboardingCompleted = new EventMetricType({
    category: "main",
    name: "onboarding_completed",
    sendInPings: ["main-events"],
    lifetime: "ping",
    disabled: false,
}, []);

/**
 * Event sent on video plays
 *
 * Generated from `main.video_played`.
 */
export const videoPlayed = new EventMetricType<{
    videos_played?: number,
}>({
    category: "main",
    name: "video_played",
    sendInPings: ["main-events"],
    lifetime: "ping",
    disabled: false,
}, ["videos_played"]);

/**
 * Event sent on injected regret button clicks
 *
 * Generated from `main.regret_action`.
 */
export const regretAction = new EventMetricType<{
    feedback_type?: string,
    video_data_id?: string,
}>({
    category: "main",
    name: "regret_action",
    sendInPings: ["main-events"],
    lifetime: "ping",
    disabled: false,
}, ["feedback_type", "video_data_id"]);

/**
 * Event sent on native YT UI clicks
 *
 * Generated from `main.native_ui_interaction`.
 */
export const nativeUiInteraction = new EventMetricType<{
    feedback_type?: string,
    video_data_id?: string,
}>({
    category: "main",
    name: "native_ui_interaction",
    sendInPings: ["main-events"],
    lifetime: "ping",
    disabled: false,
}, ["feedback_type", "video_data_id"]);

/**
 * Event sent on YT video recommendation capture
 *
 * Generated from `main.video_recommended`.
 */
export const videoRecommended = new EventMetricType<{
    recommendation_type?: string,
    video_data_id?: string,
}>({
    category: "main",
    name: "video_recommended",
    sendInPings: ["main-events"],
    lifetime: "ping",
    disabled: false,
}, ["recommendation_type", "video_data_id"]);