// source/telemetry/generated/metadata.ts
/* This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/. */

// AUTOGENERATED BY glean_parser. DO NOT EDIT. DO NOT COMMIT.

import StringMetricType from "@mozilla/glean/private/metrics/string";

/**
 * Feedback UI variant
 *
 * Generated from `metadata.feedback_ui_variant`.
 */
export const feedbackUiVariant = new StringMetricType({
    category: "metadata",
    name: "feedback_ui_variant",
    sendInPings: ["main-events", "video-data"],
    lifetime: "user",
    disabled: false,
});

/**
 * Experiment arm
 *
 * Generated from `metadata.experiment_arm`.
 */
export const experimentArm = new StringMetricType({
    category: "metadata",
    name: "experiment_arm",
    sendInPings: ["main-events", "video-data"],
    lifetime: "user",
    disabled: false,
});

/**
 * Installation ID
 *
 * Generated from `metadata.installation_id`.
 */
export const installationId = new StringMetricType({
    category: "metadata",
    name: "installation_id",
    sendInPings: ["main-events", "video-data"],
    lifetime: "user",
    disabled: false,
});