// source/telemetry/generated/videoData.ts
/* This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/. */

// AUTOGENERATED BY glean_parser. DO NOT EDIT. DO NOT COMMIT.

import QuantityMetricType from "@mozilla/glean/private/metrics/quantity";
import StringMetricType from "@mozilla/glean/private/metrics/string";
import TextMetricType from "@mozilla/glean/private/metrics/text";

/**
 * Video data UUID
 *
 * Generated from `video_data.uuid`.
 */
export const uuid = new StringMetricType({
    category: "video_data",
    name: "uuid",
    sendInPings: ["video-data"],
    lifetime: "ping",
    disabled: false,
});

/**
 * Video Id
 *
 * Generated from `video_data.id`.
 */
export const id = new StringMetricType({
    category: "video_data",
    name: "id",
    sendInPings: ["video-data"],
    lifetime: "ping",
    disabled: false,
});

/**
 * Video title
 *
 * Generated from `video_data.title`.
 */
export const title = new StringMetricType({
    category: "video_data",
    name: "title",
    sendInPings: ["video-data"],
    lifetime: "ping",
    disabled: false,
});

/**
 * Video view count
 *
 * Generated from `video_data.view_count`.
 */
export const viewCount = new StringMetricType({
    category: "video_data",
    name: "view_count",
    sendInPings: ["video-data"],
    lifetime: "ping",
    disabled: false,
});

/**
 * Video duration
 *
 * Generated from `video_data.duration`.
 */
export const duration = new QuantityMetricType({
    category: "video_data",
    name: "duration",
    sendInPings: ["video-data"],
    lifetime: "ping",
    disabled: false,
});

/**
 * Video description
 *
 * Generated from `video_data.description`.
 */
export const description = new TextMetricType({
    category: "video_data",
    name: "description",
    sendInPings: ["video-data"],
    lifetime: "ping",
    disabled: false,
});

/**
 * Video posted date
 *
 * Generated from `video_data.posted_date`.
 */
export const postedDate = new StringMetricType({
    category: "video_data",
    name: "posted_date",
    sendInPings: ["video-data"],
    lifetime: "ping",
    disabled: false,
});

/**
 * Owner channel Id
 *
 * Generated from `video_data.channel_id`.
 */
export const channelId = new StringMetricType({
    category: "video_data",
    name: "channel_id",
    sendInPings: ["video-data"],
    lifetime: "ping",
    disabled: false,
});