// source/telemetry/generated/regretDetails.ts
/* This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/. */

// AUTOGENERATED BY glean_parser. DO NOT EDIT. DO NOT COMMIT.

import StringMetricType from "@mozilla/glean/private/metrics/string";
import TextMetricType from "@mozilla/glean/private/metrics/text";

/**
 * Video UUID (not to be confused with YT video id)
 *
 * Generated from `regret_details.video_data_id`.
 */
export const videoDataId = new StringMetricType({
    category: "regret_details",
    name: "video_data_id",
    sendInPings: ["regret-details"],
    lifetime: "ping",
    disabled: false,
});

/**
 * Regret feedback text
 *
 * Generated from `regret_details.feedback_text`.
 */
export const feedbackText = new TextMetricType({
    category: "regret_details",
    name: "feedback_text",
    sendInPings: ["regret-details"],
    lifetime: "ping",
    disabled: false,
});