// source/telemetry/generated/pings.ts
/* This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/. */

// AUTOGENERATED BY glean_parser. DO NOT EDIT. DO NOT COMMIT.

import PingType from "@mozilla/glean/private/ping";

/**
 * A ping to record various extension events
 *
 * Generated from `main-events`.
 */
export const mainEvents = new PingType({
    includeClientId: true,
    sendIfEmpty: false,
    name: "main-events",
    reasonCodes: [],
});

/**
 * Ping sent once a user submits regret feedback details
 *
 * Generated from `regret-details`.
 */
export const regretDetails = new PingType({
    includeClientId: true,
    sendIfEmpty: false,
    name: "regret-details",
    reasonCodes: [],
});

/**
 * A ping to record video data
 *
 * Generated from `video-data`.
 */
export const videoData = new PingType({
    includeClientId: false,
    sendIfEmpty: false,
    name: "video-data",
    reasonCodes: [],
});